payment gateway integration

+++ restr/models.py
@@ -6,7 +6,7 @@ from django.contrib.auth.models import User
 
 class Menu(models.Model):
     items = models.CharField(max_length=200)
-    price = models.FloatField(max_length=100)
+    price = models.IntegerField(max_length=100)
 
     def __str__(self):
         return self.items
